fix: move LTX2 dummy generators to input_generators.py and remove rebase duplicates
